.uppdev
git-svn-id: svn://ultimatepp.org/upp/trunk@6255 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppdev/SDL20/SDLDraw.cpp
@@ -53,7 +53,7 @@ ImageSysData::~ImageSysData()
 
 struct ImageSysDataMaker : LRUCache<ImageSysData, int64>::Maker {
 	Image         img;
-	int           wserial;
+	int64         wserial;
 	SDL_Renderer *renderer;
 
 	virtual int64  Key() const                      { return img.GetSerialId(); } // TODO! add ren_serial
